model code restructure
